Move flawless victory to common
