+++ front/src/controls/flawless-victory.scss
@@ -1,0 +1,276 @@
+@use "sass:math";
+
+.flawless-victory {
+  $particles: 50;
+  $width: 500;
+  $height: 500;
+
+  // Create the explosion...
+  $box-shadow: ();
+  $box-shadow2: ();
+  @for $i from 0 through $particles {
+    $box-shadow:
+      $box-shadow,
+      math.div(math.random($width)-$width, 2) +
+        px
+        math.div(math.random($height)-$height, 1.5) +
+        px
+        hsl(math.random(360), 100%, 50%);
+    $box-shadow2:
+      $box-shadow2,
+      0 0 #fff;
+  }
+  @mixin keyframes($animationName) {
+    @-webkit-keyframes #{$animationName} {
+      @content;
+    }
+
+    @-moz-keyframes #{$animationName} {
+      @content;
+    }
+
+    @-o-keyframes #{$animationName} {
+      @content;
+    }
+
+    @-ms-keyframes #{$animationName} {
+      @content;
+    }
+
+    @keyframes #{$animationName} {
+      @content;
+    }
+  }
+
+  @mixin animation-delay($settings) {
+    -moz-animation-delay: $settings;
+    -webkit-animation-delay: $settings;
+    -o-animation-delay: $settings;
+    -ms-animation-delay: $settings;
+    animation-delay: $settings;
+  }
+
+  @mixin animation-duration($settings) {
+    -moz-animation-duration: $settings;
+    -webkit-animation-duration: $settings;
+    -o-animation-duration: $settings;
+    -ms-animation-duration: $settings;
+    animation-duration: $settings;
+  }
+
+  @mixin animation($settings) {
+    -moz-animation: $settings;
+    -webkit-animation: $settings;
+    -o-animation: $settings;
+    -ms-animation: $settings;
+    animation: $settings;
+  }
+
+  @mixin transform($settings) {
+    transform: $settings;
+    -moz-transform: $settings;
+    -webkit-transform: $settings;
+    -o-transform: $settings;
+    -ms-transform: $settings;
+  }
+
+  .pyro > .before,
+  .pyro > .after {
+    position: absolute;
+    width: 5px;
+    height: 5px;
+    top: 0;
+    border-radius: 50%;
+    box-shadow: $box-shadow2;
+    @include animation(
+      (
+        1s bang ease-out infinite backwards,
+        1s gravity ease-in infinite backwards,
+        5s position linear infinite backwards
+      )
+    );
+  }
+
+  .pyro > .after {
+    @include animation-delay((1.25s, 1.25s, 1.25s));
+    @include animation-duration((1.25s, 1.25s, 6.25s));
+  }
+
+  @include keyframes(bang) {
+    to {
+      box-shadow: $box-shadow;
+    }
+  }
+
+  @include keyframes(gravity) {
+    to {
+      @include transform(translateY(200px));
+      opacity: 0;
+    }
+  }
+
+  @include keyframes(position) {
+    0%,
+    19.9% {
+      margin-top: 10%;
+      margin-left: 40%;
+    }
+    20%,
+    39.9% {
+      margin-top: 40%;
+      margin-left: 30%;
+    }
+    40%,
+    59.9% {
+      margin-top: 20%;
+      margin-left: 70%;
+    }
+    60%,
+    79.9% {
+      margin-top: 30%;
+      margin-left: 20%;
+    }
+    80%,
+    99.9% {
+      margin-top: 30%;
+      margin-left: 80%;
+    }
+  }
+
+  .star-decoration {
+    position: absolute;
+    width: 40px;
+    height: 40px;
+    clip-path: polygon(50% 0%, 61% 35%, 98% 35%, 68% 57%, 79% 91%, 50% 70%, 21% 91%, 32% 57%, 2% 35%, 39% 35%);
+    z-index: -1;
+    animation: rotateStar 10s linear infinite;
+  }
+
+  @keyframes rotateStar {
+    0% {
+      transform: rotate(0deg);
+    }
+    100% {
+      transform: rotate(360deg);
+    }
+  }
+
+  .star-1 {
+    top: 40px;
+    left: 40px;
+    background: linear-gradient(135deg, #ffc107, #ff9800);
+  }
+
+  .star-2 {
+    bottom: 40px;
+    right: 40px;
+    background: linear-gradient(135deg, #ffc107, #ff9800);
+  }
+
+  .star-3 {
+    top: 60%;
+    left: 30px;
+    background: linear-gradient(135deg, #4299e1, #63b3ed);
+    width: 30px;
+    height: 30px;
+    animation-duration: 8s;
+    animation-direction: reverse;
+  }
+
+  .star-4 {
+    top: 20%;
+    right: 50px;
+    background: linear-gradient(135deg, #48bb78, #38a169);
+    width: 25px;
+    height: 25px;
+    animation-duration: 12s;
+  }
+
+  .image-container {
+    text-align: center;
+    animation: floatImage 3s ease-in-out infinite;
+    .reward-image {
+      width: 300px;
+      height: auto;
+      margin: auto;
+      border-radius: 15px;
+      box-shadow: 0 10px 20px rgba(0, 0, 0, 0.2);
+    }
+  }
+
+  @keyframes floatImage {
+    0% {
+      transform: translateY(0px);
+    }
+    50% {
+      transform: translateY(-10px);
+    }
+    100% {
+      transform: translateY(0px);
+    }
+  }
+
+  .score-display {
+    font-size: 24px;
+    font-weight: bold;
+    color: #ed8936;
+    margin: 20px 0;
+  }
+  .message {
+    font-size: 18px;
+    color: #2d3748;
+    margin: 15px 0;
+    line-height: 1.5;
+  }
+  .buttons-container {
+    display: flex;
+    justify-content: center;
+    gap: 20px;
+    margin-top: 30px;
+  }
+
+  .button {
+    background: linear-gradient(135deg, #6a5acd, #4834d4);
+    color: white;
+    border: none;
+    border-radius: 12px;
+    padding: 12px 18px;
+    font-size: 16px;
+    font-weight: bold;
+    cursor: pointer;
+    box-shadow: 0 4px 6px rgba(106, 90, 205, 0.3);
+    transition: all 0.2s ease;
+    position: relative;
+    overflow: hidden;
+  }
+
+  .button-restart {
+    background: linear-gradient(135deg, #f56565, #e53e3e);
+  }
+
+  .button::after {
+    content: "";
+    position: absolute;
+    top: 0;
+    left: 0;
+    right: 0;
+    bottom: 0;
+    background: linear-gradient(rgba(255, 255, 255, 0.1), rgba(255, 255, 255, 0));
+    opacity: 0;
+    transition: opacity 0.2s ease;
+  }
+
+  .button:hover {
+    transform: translateY(-3px);
+    box-shadow: 0 6px 12px rgba(106, 90, 205, 0.4);
+  }
+
+  .button:hover::after {
+    opacity: 1;
+  }
+
+  .button:active {
+    transform: translateY(0);
+    box-shadow: 0 2px 3px rgba(106, 90, 205, 0.3);
+  }
+}
